monitor-flask-service/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.9
python-dotenv==1.1.0
matplotlib==3.10.1
plotly
requests
requests-toolbelt